+++ CHANGELOG.md
@@ -19,6 +19,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * `RandHashProj` - Algorithm to implement the "hashing trick" in an easy-to-use, memory efficient way.
 * `turbo` color mapping (primarily for depth images)
 * `nyuv2` dataset
+* `TensorBoardAddDepth` callback.
 
 ## Fixed
 * Prevent docker container from exitting when vnc disconnects/crashes.